% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchPlotGroups.R
\name{batchPlot.groups}
\alias{batchPlot.groups}
\title{Plot a Set of Curve Groups from a List of Data Groups}
\usage{
batchPlot.groups(data, extract = identity,
  colors = colors.distinct(length(data)), ...)
}
\arguments{
\item{data}{the data object, could be a list of lists or anything}

\item{extract}{a function which extracts a list of data elements from each
element of \code{data}. This list will then be passed on to
\code{\link{batchPlot.list}}.}

\item{colors}{the colors to be used for the plot}

\item{...}{Arguments passed on to \code{batchPlot.list}
\describe{
  \item{xfun}{a function which receives an element from the \code{data} list
and extracts a vector of \code{x}-coordinates from it}
  \item{yfun}{a function which receives an element from the \code{data} list
and extracts a vector of \code{y}-coordinates from it to be plotted as
points, or \code{NULL} if no points should be plotted (see \code{plotXY})}
  \item{ffun}{a function which receives an element from the \code{data} list
and the corresponding extracted \code{x}-coordinates and extracts or
computes a vector of \code{y}-coordinates from it to be plotted as line, or
\code{NULL} if no points should be plotted (see \code{plotXF})}
  \item{plotXY}{should the \code{x-y} points be plotted (if \code{yfun} is not
\code{NULL})}
  \item{widthXY}{the line width for points to be plotted (only considered if
\code{plotXY} is \code{TRUE} and \code{yfun} is not \code{NULL})}
  \item{plotXF}{should the \code{x-y} lines be plotted (if \code{ffun} is not
\code{NULL})}
  \item{widthXF}{the line width for lines to be plotted (only considered if
\code{plotXF} is \code{TRUE} and \code{ffun} is not \code{NULL})}
  \item{names}{the names of the lines to be printed in the legend, or
\code{NULL} if no legend should be plotted}
  \item{xlab}{the label for the x-axis}
  \item{ylab}{the label for the y-axis}
  \item{legend}{a list of additional parameters to be passed to
\code{\link[graphics]{legend}}, or \code{NULL} to use the default
parameters}
}}
}
\description{
A simple utility method for visualizing a list of data groups.
}
\examples{
library(plotteR)

# set a random seed for replicability
set.seed(2677);

# the three base functions with the mean parameter values
f <- list(
  list(f=function(x, par) par[1] + par[2]*x + par[3]*x*x + par[4]*x*x*x,
       m=c(1, -0.2, 0.75, -0.3)),
  list(f = function(x, par) par[1] * exp(par[2] - x),
       m=c(0.1, 3)),
  list(f=function(x, par) par[1] + par[2]*sin(par[3]*x),
       m=c(0, 1, 3)));

# create the three example data sets
examples <- lapply(X=f, FUN=function(example) {
# for each example function, plot 4 to 50 instances
  lapply(seq_len(runif(n=1, min=4, max=50)),
# for each instance
         FUN=function(i) {
# randomly choose the x-coordinates
           x <- runif(n=as.integer(round(runif(n=1, min=10, max=200))),
                      min=0, max=3);
           m <- example$m;
# pick parameters which are normally distributed around the suggestion
           par <- rnorm(n=length(m), mean=m, s=0.1*abs(m));
# and construct a function
           fff <- function(x) example$f(x, par);
# and pass this function as result example together with the x values
           list(x=x, f=fff)
         })
});

# plot the original data
batchPlot.groups(examples,
                 names=c("f1", "f2", "f3"),
                 ffun = function(l, x) l$f(x),
                 main="Several Groups of Functions",
                 plotXY=FALSE, plotXF=TRUE,
                 legend=list(x="bottom", horiz=TRUE));
}
