% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distinctColors.R
\name{colors.distinct}
\alias{colors.distinct}
\title{Get \code{n} Distinct Colors}
\usage{
colors.distinct(n)
}
\arguments{
\item{n}{the number of distinct colors we want}
}
\value{
a vector of \code{n} distinct colors
}
\description{
Get a list of \code{n} distinct colors.
This is currently just a very preliminary method, not beautiful and far from
optimal. For instance, if more colors are required than what the internally
pre-defined palettes can do, we try to uniformly sample from the RGB space
with the goal to get maximally distant colors.
}
