% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ertFun.R
\name{func.ert}
\alias{func.ert}
\title{Create an ERT Function}
\usage{
func.ert(x, goal.dim = 2L, time.dim = 1L, comparator = `<=`,
  time.type = as.integer, goal.type = identity, goal.min = goal.type(NA),
  goal.max = goal.type(NA), time.max = time.type(NA),
  extract.run = identity)
}
\arguments{
\item{x}{the data, maybe a list of matrices or a list of vectors/lists, where
each element has at least the \code{goal.dim} and \code{time.dim}
dimension.}

\item{goal.dim}{the dimension where where the goal values can be found}

\item{time.dim}{the dimension where the time values can be found}

\item{comparator}{the comparator, usually \code{`<=`} or \code{`>=`}}

\item{time.type}{the type function the time dimension, should be
\code{as.integer} or \code{identity}}

\item{goal.type}{the type function the goal dimension, should be
\code{as.integer} or \code{identity}}

\item{goal.min}{the minimum goal value to use, or \code{NA} to use the
minimal occuring value}

\item{goal.max}{the maximum goal value to use, or \code{NA} to use the
maximally occuring one}

\item{time.max}{the maximum time value to be used, or \code{NA} to pick the
maximum time value of each run separately}

\item{extract.run}{a function which can be used to extract the single runs
from \code{x}, e.g., \code{identity}}
}
\value{
a \code{list(x=c(...), y=c(...))} where the two members \code{x} and
  \code{y} will be vectors of corresponding coordinates
}
\description{
Create an Expected Running Time (ERT) function, i.e., a list
  with two membery \code{x} and \code{y} that stand for the coordinates of
  the function.
}
