% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distinctLineTypes.R
\name{lineTypes.distinct}
\alias{lineTypes.distinct}
\title{Create a List of Distinct Line Tpes of Length \code{n}}
\usage{
lineTypes.distinct(n)
}
\arguments{
\item{n}{the number of different line types to generate}
}
\value{
the vector of line types
}
\description{
Create a vector of line types of a given length \code{n} that can
  be used in plots. The line types are chosen to be as different as possible,
  but sometimes the same line type may occur, since only a given number of
  different line types exist in R.
}
