% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotColors.R
\name{plot.colors}
\alias{plot.colors}
\title{Plot a Vector of Colors so that we can see if they are unique}
\usage{
\method{plot}{colors}(x, ...)
}
\arguments{
\item{x}{the colors to plot}

\item{...}{the parameters passed to \code{pie} or \code{plot}}
}
\description{
Take a vector of colors \code{x} and plot them in a way so that
  each pair of colors occurs once. This allows us to see if they are unique
  and distinct.
}
