% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecdf.R
\name{plot.ecdf}
\alias{plot.ecdf}
\title{Plot ECDF Functions}
\usage{
\method{plot}{ecdf}(x, goal = 0, goal.dim = 2L, time.dim = 1L,
  time.type = as.integer, time.min = time.type(1L),
  time.max = time.type(NA_integer_), goal.min = 0, goal.max = 1,
  extract.runs = identity, comparator = `<=`, extract.run = identity,
  lineTypeFun = lineTypes.distinct, colorFun = colors.distinct,
  names = NULL, goal.markers = c(0.2, 0.4, 0.6, 0.8), time.markers = NULL,
  ...)
}
\arguments{
\item{x}{the data, maybe a list of [lists of matrices or a list of
vectors/lists], where each element has at least the \code{goal.dim} and
\code{time.dim} dimension.}

\item{goal}{the goal value, i.e., the value which must be reached to be
counted as success}

\item{goal.dim}{the dimension where where the goal values can be found}

\item{time.dim}{the dimension where the time values can be found}

\item{time.type}{the type function the time dimension, should be
\code{as.integer} or \code{identity}}

\item{time.min}{the minimum time value to be used for the diagram, or
\code{NA} to use the smallest time value in any run}

\item{time.max}{the maximum time value to be used, or \code{NA} to pick the
maximum time value of any run}

\item{goal.min}{the minimum goal value}

\item{goal.max}{the maximum goal value}

\item{extract.runs}{a function which can be used to extract the run sets from
\code{x}, e.g., \code{identity}}

\item{comparator}{the comparator, usually \code{`<=`} or \code{`>=`}}

\item{extract.run}{a function which can be used to extract the single runs
from the run sets, e.g., \code{identity}}

\item{lineTypeFun}{the line type function, a function returning the line
types to use for a provided number of inputs}

\item{colorFun}{the colors function, a function returning the color list to
use for a provided number of inputs}

\item{names}{the legend}

\item{goal.markers}{markers for the goal values}

\item{time.markers}{markers for the time values}

\item{...}{Arguments passed on to \code{graphics::plot}
\describe{
  \item{x}{the coordinates of points in the plot. Alternatively, a
      single plotting structure, function or \emph{any \R object with a
        \code{plot} method} can be provided.}
  \item{y}{the y coordinates of points in the plot, \emph{optional}
    if \code{x} is an appropriate structure.}
}}
}
\description{
Plot the ECDF functions for a couple of different datasets.
}
