% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ert.R
\name{plot.ert}
\alias{plot.ert}
\title{Plot ERT Functions}
\usage{
\method{plot}{ert}(x, goal.dim = 2L, time.dim = 1L,
  time.type = as.integer, goal.type = as.numeric,
  goal.min = goal.type(NA_real_), goal.max = goal.type(NA_real_),
  time.max = time.type(NA_real_), extract.run = identity,
  extract.runs = identity, comparator = `<=`,
  lineTypeFun = lineTypes.distinct, colorFun = colors.distinct,
  names = NULL, goal.markers = c(0.2, 0.4, 0.6, 0.8), time.markers = NULL,
  ...)
}
\arguments{
\item{x}{the data, maybe a list of [lists of matrices or a list of
vectors/lists], where each element has at least the \code{goal.dim} and
\code{time.dim} dimension.}

\item{goal.dim}{the dimension where where the goal values can be found}

\item{time.dim}{the dimension where the time values can be found}

\item{time.type}{the type function the time dimension, should be
\code{as.integer} or \code{as.numeric}}

\item{goal.type}{the type function the goal dimension, should be
\code{as.integer} or \code{as.numeric}}

\item{goal.min}{the minimum goal value or \code{NA} to pick the minimum time
value}

\item{goal.max}{the maximum goal value or \code{NA} to pick the minimum time
value}

\item{time.max}{the maximum time value to be used, or \code{NA} to pick the
maximum time value of any run}

\item{extract.run}{a function which can be used to extract the single runs
from the run sets, e.g., \code{identity}}

\item{extract.runs}{a function which can be used to extract the run sets from
\code{x}, e.g., \code{identity}}

\item{comparator}{the comparator, usually \code{`<=`} or \code{`>=`}}

\item{lineTypeFun}{the line type function, a function returning the line
types to use for a provided number of inputs}

\item{colorFun}{the colors function, a function returning the color list to
use for a provided number of inputs}

\item{names}{the legend}

\item{goal.markers}{markers for the goal values}

\item{time.markers}{markers for the time values}

\item{...}{Arguments passed on to \code{graphics::plot}
\describe{
  \item{x}{the coordinates of points in the plot. Alternatively, a
      single plotting structure, function or \emph{any \R object with a
        \code{plot} method} can be provided.}
  \item{y}{the y coordinates of points in the plot, \emph{optional}
    if \code{x} is an appropriate structure.}
}}
}
\description{
Plot the Expected Running Time (ERT) functions for a couple of
  different datasets.
}
