% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gantt.R
\name{plot.gantt}
\alias{plot.gantt}
\title{Plot a Gantt Chart}
\usage{
\method{plot}{gantt}(x, xlab = "Time", ylab = "Machine", prefix.job = "J",
  prefix.machine = "M", color.fun = colors.distinct, ...)
}
\arguments{
\item{x}{the x list, a list of lists}

\item{xlab}{the \code{x} label}

\item{ylab}{the \code{y} label}

\item{prefix.job}{the job name prefix}

\item{prefix.machine}{the machine name prefix}

\item{color.fun}{a function returning a color for a job}

\item{...}{Arguments passed on to \code{graphics::plot}
\describe{
  \item{}{}
}}
}
\description{
Plot a Gantt chart based on a list \code{x} of lists of data. The
  list contains one list for each machine. The first machine at index
  \code{1} will be refered to as \code{M0}, the second machine as \code{M1}
  and so on. Each machine list, in turn, is a list of lists. Each element has
  the form \code{list(job=, start=, end=)}, with elements denoting the job
  name, start, and end time, respectively.
}
\examples{
#
if(!(require("plotteR"))){
  if(!(require("devtools"))){
    install.packages("devtools");
    library("devtools");
  };
  devtools::install_github("thomasWeise/plotteR");
  library("plotteR");
};

plot.gantt(list(
  list( list(job=11L,start=0L,end=68L),
        list(job=0L,start=83L,end=155L),
        list(job=8L,start=155L,end=194L),
        list(job=14L,start=250L,end=297L),
        list(job=1L,start=297L,end=320L),
        list(job=9L,start=320L,end=325L),
        list(job=2L,start=325L,end=380L),
        list(job=5L,start=380L,end=429L),
        list(job=3L,start=457L,end=548L),
        list(job=13L,start=548L,end=643L),
        list(job=6L,start=643L,end=718L),
        list(job=12L,start=718L,end=798L),
        list(job=10L,start=798L,end=822L),
        list(job=4L,start=822L,end=865L),
        list(job=7L,start=865L,end=937L)
  ),
  list( list(job=3L,start=0L,end=68L),
        list(job=2L,start=68L,end=118L),
        list(job=12L,start=211L,end=271L),
        list(job=6L,start=315L,end=405L),
        list(job=1L,start=405L,end=490L),
        list(job=13L,start=490L,end=513L),
        list(job=9L,start=513L,end=601L),
        list(job=11L,start=601L,end=619L),
        list(job=4L,start=619L,end=689L),
        list(job=5L,start=689L,end=758L),
        list(job=14L,start=758L,end=786L),
        list(job=0L,start=786L,end=812L),
        list(job=7L,start=812L,end=821L),
        list(job=8L,start=821L,end=901L),
        list(job=10L,start=901L,end=950L)
  ),
  list( list(job=14L,start=0L,end=24L),
        list(job=13L,start=45L,end=137L),
        list(job=6L,start=137L,end=228L),
        list(job=4L,start=228L,end=262L),
        list(job=11L,start=314L,end=391L),
        list(job=2L,start=391L,end=452L),
        list(job=0L,start=452L,end=490L),
        list(job=12L,start=490L,end=530L),
        list(job=7L,start=530L,end=621L),
        list(job=10L,start=622L,end=640L),
        list(job=8L,start=640L,end=703L),
        list(job=9L,start=703L,end=771L),
        list(job=1L,start=771L,end=838L),
        list(job=3L,start=844L,end=916L),
        list(job=5L,start=960L,end=969L)
  ),
  list( list(job=2L,start=118L,end=211L),
        list(job=1L,start=211L,end=254L),
        list(job=11L,start=254L,end=314L),
        list(job=12L,start=314L,end=370L),
        list(job=9L,start=370L,end=420L),
        list(job=7L,start=420L,end=451L),
        list(job=13L,start=473L,end=479L),
        list(job=10L,start=479L,end=570L),
        list(job=8L,start=570L,end=622L),
        list(job=14L,start=622L,end=699L),
        list(job=3L,start=699L,end=795L),
        list(job=4L,start=795L,end=807L),
        list(job=6L,start=807L,end=824L),
        list(job=0L,start=824L,end=843L),
        list(job=5L,start=933L,end=960L)
    ),
  list( list(job=8L,start=0L,end=35L),
        list(job=13L,start=35L,end=45L),
        list(job=3L,start=135L,end=234L),
        list(job=0L,start=234L,end=264L),
        list(job=2L,start=291L,end=298L),
        list(job=5L,start=429L,end=488L),
        list(job=1L,start=490L,end=529L),
        list(job=6L,start=529L,end=537L),
        list(job=10L,start=570L,end=603L),
        list(job=7L,start=621L,end=659L),
        list(job=14L,start=786L,end=834L),
        list(job=9L,start=834L,end=887L),
        list(job=12L,start=887L,end=938L),
        list(job=11L,start=938L,end=956L),
        list(job=4L,start=956L,end=986L)
  ),
  list( list(job=14L,start=24L,end=53L),
        list(job=13L,start=137L,end=160L),
        list(job=2L,start=211L,end=291L),
        list(job=7L,start=291L,end=389L),
        list(job=3L,start=389L,end=457L),
        list(job=8L,start=457L,end=482L),
        list(job=11L,start=482L,end=497L),
        list(job=5L,start=497L,end=569L),
        list(job=4L,start=569L,end=576L),
        list(job=1L,start=576L,end=589L),
        list(job=10L,start=603L,end=622L),
        list(job=0L,start=622L,end=720L),
        list(job=6L,start=720L,end=757L),
        list(job=9L,start=771L,end=795L),
        list(job=12L,start=798L,end=887L)
  ),
  list( list(job=1L,start=0L,end=19L),
        list(job=6L,start=19L,end=90L),
        list(job=0L,start=155L,end=229L),
        list(job=3L,start=234L,end=294L),
        list(job=13L,start=294L,end=340L),
        list(job=12L,start=370L,end=461L),
        list(job=14L,start=461L,end=538L),
        list(job=2L,start=538L,end=595L),
        list(job=9L,start=601L,end=621L),
        list(job=10L,start=640L,end=678L),
        list(job=4L,start=689L,end=763L),
        list(job=8L,start=763L,end=798L),
        list(job=11L,start=798L,end=888L),
        list(job=5L,start=888L,end=933L),
        list(job=7L,start=937L,end=986L)
  ),
  list( list(job=0L,start=0L,end=8L),
        list(job=4L,start=8L,end=92L),
        list(job=3L,start=92L,end=135L),
        list(job=7L,start=135L,end=225L),
        list(job=10L,start=225L,end=324L),
        list(job=11L,start=391L,end=401L),
        list(job=13L,start=401L,end=473L),
        list(job=8L,start=482L,end=529L),
        list(job=6L,start=537L,end=587L),
        list(job=12L,start=639L,end=711L),
        list(job=14L,start=711L,end=719L),
        list(job=5L,start=758L,end=857L),
        list(job=2L,start=857L,end=903L),
        list(job=1L,start=903L,end=912L),
        list(job=9L,start=912L,end=969L)
  ),
  list( list(job=5L,start=0L,end=60L),
        list(job=7L,start=60L,end=122L),
        list(job=1L,start=122L,end=195L),
        list(job=14L,start=195L,end=250L),
        list(job=6L,start=250L,end=315L),
        list(job=0L,start=315L,end=358L),
        list(job=13L,start=358L,end=398L),
        list(job=11L,start=401L,end=461L),
        list(job=9L,start=461L,end=513L),
        list(job=4L,start=513L,end=553L),
        list(job=12L,start=553L,end=639L),
        list(job=3L,start=639L,end=650L),
        list(job=2L,start=650L,end=722L),
        list(job=8L,start=722L,end=743L),
        list(job=10L,start=950L,end=959L)
  ),
  list( list(job=0L,start=8L,end=83L),
        list(job=14L,start=83L,end=132L),
        list(job=12L,start=132L,end=211L),
        list(job=9L,start=211L,end=269L),
        list(job=8L,start=269L,end=343L),
        list(job=6L,start=405L,end=503L),
        list(job=11L,start=503L,end=575L),
        list(job=5L,start=575L,end=638L),
        list(job=7L,start=659L,end=731L),
        list(job=1L,start=731L,end=757L),
        list(job=2L,start=757L,end=799L),
        list(job=13L,start=799L,end=833L),
        list(job=3L,start=833L,end=844L),
        list(job=10L,start=844L,end=879L),
        list(job=4L,start=879L,end=948L)
  )
), main="la24, makespan: 986", prefix.job="");
}
