% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotsArrange.R
\name{plots.arrange}
\alias{plots.arrange}
\title{Find a Proper Numbers of Rows and Columns for \code{n} Plots}
\usage{
plots.arrange(n)
}
\arguments{
\item{n}{the number of plots}
}
\value{
a vector \code{c(rows, columns)} with the number of rows and columns in which these plots can be arranged.
}
\description{
Given a number \code{n} of plots, find proper numbers of rows and columns in which these plots can be arranged.
}
