% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goldenRatio.R
\name{plots.pleasing.size}
\alias{plots.pleasing.size}
\title{Compute a Pleasing Plot Size}
\usage{
plots.pleasing.size(width = NULL, height = NULL,
  paperWidthFrac = NULL, paperHeightFrac = NULL, portrait = FALSE,
  boundToPaper = TRUE)
}
\arguments{
\item{width}{the optional width of the plot}

\item{height}{the optional height of the plot}

\item{paperWidthFrac}{optional, a fraction of the default paper width}

\item{paperHeightFrac}{optional, a fraction of the default paper height}

\item{portrait}{optional, should we use portrait orientation or letter
orientation({@code FALSE})}

\item{boundToPaper}{should we bound the result dimensions to the paper
size?}
}
\value{
a \code{list(width, height)} of reasonable figure dimensions
}
\description{
Return a pleasing size for plots, based on the {@code width} and
  {@code height}, if provided.
}
\examples{
plots.pleasing.size(width=3)
# $width=3, $height=1.854102
plots.pleasing.size(height=3)
# $width=4.854102, $height=3
plots.pleasing.size()
# $width=8.3, $height=5.129682
}
