\name{plotter-package}
\alias{plotter-package}
\alias{plotter}
\docType{package}
\title{
\packageTitle{plotter}
}
\description{
\packageDescription{plotter}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{plotter}
\packageIndices{plotter}
How to use the package...
}
\author{
\packageAuthor{plotter}

Maintainer: \packageMaintainer{plotter}
}

\examples{
~~ simple examples of the most important functions ~~
}
