% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotter.R
\name{plotter}
\alias{plotter}
\title{Saves plot data to disk}
\usage{
plotter(inp_data = NULL, graphType = NULL, graphOptions = list())
}
\arguments{
\item{inp_data}{data to be plotted}

\item{graphType}{type of graph to be plotted}

\item{graphOptions}{options for the selected graph}
}
\value{
nothing, data is stored to disk
}
\description{
Saves plot data to disk
}
