% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgbScaleLuminosity.R
\name{rgb.scale.luminosity}
\alias{rgb.scale.luminosity}
\title{Scale the Luminosity of a Color}
\usage{
rgb.scale.luminosity(r, g, b, scale, limit = 255)
}
\arguments{
\item{r}{the red value}

\item{g}{the green value}

\item{b}{the blue value}

\item{scale}{the scale value, in \code{[-1,1]}}

\item{limit}{the maximum permitted value for each rgb-coordinate: defaults to
255}
}
\value{
a color with similar chromotographical properties but either lighter
  (scale>0) or darker (scale<0) complexion.
}
\description{
The luminosity of a color is scaled in a range of \code{[-1,1]},
  where -1 means black, 1 means white, and 0 means the color at its current
  luminosity. A value in \code{(0, 1]} will make the color brighter while
  trying to preserve its chromotographical properties. A value in \code{[-1,
  0)} will make the color darker while trying to preserve its
  chromotographical properties. See
  \url{https://stackoverflow.com/questions/6615002}
}
