% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgb2gray.R
\name{rgb2gray.luminosity}
\alias{rgb2gray.luminosity}
\title{Convert a RGB Color to Gray Scale using Luminosity Weights}
\usage{
rgb2gray.luminosity(r, g, b)
}
\arguments{
\item{r}{the red value}

\item{g}{the green value}

\item{b}{the blue value}
}
\value{
the gray scale value.
}
\description{
We apply the simple algorithm from
  \url{http://www.johndcook.com/blog/2009/08/24/algorithms-convert-color-grayscale/}
  to convert rgb values to gray-scale based on luminosity. Notice that this
  function does not care about the range of \code{r}, \code{g}, and \code{b}:
  If all of them are in \code{0..255}, then the output will also be in
  \code{[0,255]} (though not necessary integer). If they are in \code{[0,1]},
  the output is also in \code{[0,1]}. As coefficients, we use the values for
  BT.709 (\url{http://en.wikipedia.org/wiki/Rec._709}).
}
