% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distinctSymbols.R
\name{symbols.distinct}
\alias{symbols.distinct}
\title{Create a List of Distinct Symbols of Length \code{n}}
\usage{
symbols.distinct(n)
}
\arguments{
\item{n}{the number of different symbols to generate}
}
\value{
the vector of symbols
}
\description{
Create a vector of symbols of a given length \code{n} that can
  be used in plots. The symbols are chosen to be as different as possible,
  but sometimes the same symbols may occur, since only a given number of
  different symbols exist in R.
}
