% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchPlotList.R
\name{batchPlot.list}
\alias{batchPlot.list}
\title{Plot a Set of Curves from a List of Data}
\usage{
batchPlot.list(data, log = "", xfun = function(d) d$x, yfun = function(d)
  d$y, ffun = function(d, x) yfun(d), plotXY = TRUE, widthXY = 0.5,
  plotXF = TRUE, widthXF = 1.5, names = NULL,
  colors = colors.distinct(length(data)), xlab = "", ylab = "",
  legendPos = "topright")
}
\arguments{
\item{data}{the data object, could be a list of lists or anything}

\item{log}{the names of the axes (\code{x}, \code{y}) that should be
logarithmically scaled}

\item{xfun}{a function which receives an element from the \code{data} list
and extracts a vector of \code{x}-coordinates from it}

\item{yfun}{a function which receives an element from the \code{data} list
and extracts a vector of \code{y}-coordinates from it to be plotted as
points, or \code{NULL} if no points should be plotted (see \code{plotXY})}

\item{ffun}{a function which receives an element from the \code{data} list
and the corresponding extracted \code{x}-coordinates and extracts or
computes a vector of \code{y}-coordinates from it to be plotted as line, or
\code{NULL} if no points should be plotted (see \code{plotXF})}

\item{plotXY}{should the \code{x-y} points be plotted (if \code{yfun} is not
\code{NULL})}

\item{widthXY}{the line width for points to be plotted (only considered if
\code{plotXY} is \code{TRUE} and \code{yfun} is not \code{NULL})}

\item{plotXF}{should the \code{x-y} lines be plotted (if \code{ffun} is not
\code{NULL})}

\item{widthXF}{the line width for lines to be plotted (only considered if
\code{plotXF} is \code{TRUE} and \code{ffun} is not \code{NULL})}

\item{names}{the names of the lines to be printed in the legend, or
\code{NULL} if no legend should be plotted}

\item{colors}{the colors to be used for the plot}

\item{xlab}{the label for the \code{x}-axis, or the empty string if no label
is needed}

\item{ylab}{the label for the \code{y}-axis, or the empty string if no label
is needed}

\item{legendPos}{the position of the legend, if a legend should be printed
(see \code{names}), or \code{NULL} if no legend is needed}

\item{results}{an instance or list of \code{\link{RegressionResult}}}
}
\description{
A simple utility method for visualizing a list of data data.
}
