% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distinctColors.R
\name{colors.distinct}
\alias{colors.distinct}
\title{Get \code{n} Distinct Colors}
\usage{
colors.distinct(n)
}
\arguments{
\item{n}{the number of distinct colors we want}
}
\value{
a vector of \code{n} distinct colors
}
\description{
Get a list of \code{n} distinct colors.
}
