% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distinctColors.R
\name{colors.distinct}
\alias{colors.distinct}
\title{Get \code{n} Distinct Colors}
\usage{
colors.distinct(n)
}
\arguments{
\item{n}{the number of distinct colors we want}
}
\value{
a vector of \code{n} distinct colors
}
\description{
Get a list of \code{n} distinct colors.
This is currently just a very preliminary method, not beautiful and far from
optimal. For instance, if more colors are required than what the internally
pre-defined palettes can do, we try to uniformly sample from the RGB space
with the goal to get maximally distant colors.
}
\examples{
library(plotteR)

# set the random seed for reproducibility
set.seed(129855L);

x.range <- 4L;
y.range <- 5L;

# set the number of diagrams appropriately
old.par <- par(mfrow=c(y.range, x.range), mai=c(0.02, 0.02, 0.02, 0.02));

# paint the diagrams
n <- 1L;
for(y in 1L:y.range) {
  for(x in 1L:x.range) {
    plot.colors(colors.distinct(n));
    n <- n + 1L;
  }
}

# reset the parameters
invisible(par(old.par));
}
