% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotColors.R
\name{plot.colors}
\alias{plot.colors}
\title{Plot a Vector of Colors so that we can see if they are unique}
\usage{
\method{plot}{colors}(colors)
}
\arguments{
\item{colors}{the colors to plot}
}
\description{
Take a vector of \code{colors} and plot them in a way so that
  each pair of colors occurs once. This allows us to see if they are unique
  and distinct.
}
