% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotColors.R
\name{plot.colors}
\alias{plot.colors}
\title{Plot a Vector of Colors so that we can see if they are unique}
\usage{
\method{plot}{colors}(x, ...)
}
\arguments{
\item{x}{the colors to plot}

\item{...}{the parameters passed to \code{pie} or \code{plot}}
}
\description{
Take a vector of colors \code{x} and plot them in a way so that
  each pair of colors occurs once. This allows us to see if they are unique
  and distinct.
}
\examples{
library(plotteR)

# set the random seed for reproducibility
set.seed(129855L);

if(!exists("x.range")) x.range <- 4L;
if(!exists("y.range")) y.range <- 5L;

# set the number of diagrams appropriately
old.par <- par(mfrow=c(y.range, x.range), mai=c(0.02, 0.02, 0.02, 0.02));

# paint the diagrams
n <- 1L;
for(y in 1L:y.range) {
  for(x in 1L:x.range) {
    plot.colors(colors.distinct(n));
    n <- n + 1L;
  }
}

# reset the parameters
invisible(par(old.par));
}
