% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gantt.R
\name{plot.gantt}
\alias{plot.gantt}
\title{Plot a Gantt Chart}
\usage{
\method{plot}{gantt}(x, xlab = "Time", ylab = "Machine",
  prefix.job = "J", prefix.machine = "M",
  color.fun = colors.distinct, print.jobs = TRUE, las = 1L,
  xaxs = "i", yaxs = "i", xrange = NULL, ...)
}
\arguments{
\item{x}{the x list, a list of lists}

\item{xlab}{the \code{x} label}

\item{ylab}{the \code{y} label}

\item{prefix.job}{the job name prefix}

\item{prefix.machine}{the machine name prefix}

\item{color.fun}{a function returning a color for a job}

\item{print.jobs}{should we print the job ids?}

\item{las}{the label orientation, see \link[graphics]{par}}

\item{xaxs}{the x-axis type, see \link[graphics]{par}}

\item{yaxs}{the y-axis type, see \link[graphics]{par}}

\item{xrange}{the range for the x-axis, \code{NULL} for auto-determine
(default)}

\item{...}{Arguments passed on to \code{graphics::plot}
\describe{
  \item{}{}
}}
}
\description{
Plot a Gantt chart based on a list \code{x} of lists of data.
  The list contains one list for each machine. The first machine at index
  \code{1} will be refered to as \code{M0}, the second machine as \code{M1}
  and so on. Each machine list, in turn, is a list of lists. Each element has
  the form \code{list(job=, start=, end=)}, with elements denoting the job
  name, start, and end time, respectively.
}
\examples{
library("plotteR");

plot.gantt(list(
  list( list(job=11L,start=0L,end=68L),
        list(job=0L,start=83L,end=155L),
        list(job=8L,start=155L,end=194L),
        list(job=14L,start=250L,end=297L),
        list(job=1L,start=297L,end=320L),
        list(job=9L,start=320L,end=340L)
  ),
  list( list(job=3L,start=0L,end=68L),
        list(job=2L,start=68L,end=118L),
        list(job=12L,start=211L,end=271L)
  ),
  list( list(job=14L,start=0L,end=24L),
        list(job=13L,start=45L,end=137L),
        list(job=6L,start=137L,end=228L),
        list(job=4L,start=228L,end=262L)
  ),
  list( list(job=2L,start=118L,end=211L),
        list(job=1L,start=211L,end=254L),
        list(job=11L,start=254L,end=314L)
  ),
  list( list(job=8L,start=0L,end=35L),
        list(job=13L,start=35L,end=45L),
        list(job=3L,start=135L,end=234L),
        list(job=0L,start=234L,end=264L),
        list(job=2L,start=291L,end=308L)
  ),
  list( list(job=14L,start=24L,end=53L),
        list(job=13L,start=137L,end=160L),
        list(job=2L,start=211L,end=291L)
  ),
  list( list(job=1L,start=0L,end=19L),
        list(job=6L,start=19L,end=90L),
        list(job=0L,start=155L,end=229L),
        list(job=3L,start=234L,end=294L),
        list(job=13L,start=294L,end=340L)
  ),
  list( list(job=0L,start=0L,end=8L),
        list(job=4L,start=8L,end=92L),
        list(job=3L,start=92L,end=135L),
        list(job=7L,start=135L,end=225L),
        list(job=10L,start=225L,end=324L)
  ),
  list( list(job=5L,start=0L,end=60L),
        list(job=7L,start=60L,end=122L),
        list(job=1L,start=122L,end=195L),
        list(job=14L,start=195L,end=250L),
        list(job=6L,start=250L,end=315L)
  ),
  list( list(job=0L,start=8L,end=83L),
        list(job=14L,start=83L,end=132L),
        list(job=12L,start=132L,end=211L),
        list(job=9L,start=211L,end=269L),
        list(job=8L,start=269L,end=343L)
  )
), main="Gantt Chart", prefix.job="");
}
