% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gantt.R
\name{plot.gantt}
\alias{plot.gantt}
\title{Plot a Gantt Chart}
\usage{
\method{plot}{gantt}(x, xlab = "Time", ylab = "Machine", prefix.job = "J",
  prefix.machine = "M", color.fun = colors.distinct, ...)
}
\arguments{
\item{x}{the x list, a list of lists}

\item{xlab}{the \code{xaxis} label}

\item{ylab}{the \code{y} label}

\item{prefix.job}{the job name prefix}

\item{prefix.machine}{the machine name prefix}

\item{color.fun}{a function returning a color for a job}
}
\description{
Plot a Gantt chart based on a list \code{x} of lists of data. The
  list contains one list for each machine. The first machine at index
  \code{1} will be refered to as \code{M0}, the second machine as \code{M1}
  and so on. Each machine list, in turn, is a list of lists. Each element has
  the form \code{list(job=, start=, end=)}, with elements denoting the job
  name, start, and end time, respectively.
}
