% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotter-package.R
\docType{package}
\name{plotter-package}
\alias{plotter}
\alias{plotter-package}
\title{plotter: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/andreyhgl/plotter}
  \item Report bugs at \url{https://github.com/andreyhgl/plotter/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
